Add files via upload
Here is the modified version of stage 1. In this version,unlike previous one, the amount of parameters {a, b, c} are different for each resolution. In general, three types of resolution has been considered {1080p, 720p, 360p}. Moreover, {r_max} is also different for each stream.
This is only a test for the stage 1. The function along with other related functions will be posted soon!